// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
// ...
      mechanism_of_injury_codes: {
        Row: {
          created_at: string | null
          moi_code_id: number
          moi_code_main: string | null
          moi_code_sub: string | null
          moi_description: string | null
          updated_at: string | null
        }
        Insert: {
          created_at?: string | null
          moi_code_id?: number
          moi_code_main?: string | null
          moi_code_sub?: string | null
          moi_description?: string | null
          updated_at?: string | null
        }
        Update: {
          created_at?: string | null
          moi_code_id?: number
          moi_code_main?: string | null
          moi_code_sub?: string | null
          moi_description?: string | null
          updated_at?: string | null
        }
        Relationships: []
      }
// ...
      sites: {
        Row: {
          city: string | null
          created_at: string | null
          employer_id: number | null
          post_code: string | null
          project_type: string | null
          site_id: number
          site_name: string | null
          state: string | null
          street_address: string | null
          supervisor_name: string | null
          supervisor_telephone: string | null
          updated_at: string | null
        }
        Insert: {
          city?: string | null
          created_at?: string | null
          employer_id?: number | null
          post_code?: string | null
          project_type?: string | null
          site_id?: number
          site_name?: string | null
          state?: string | null
          street_address?: string | null
          supervisor_name?: string | null
          supervisor_telephone?: string | null
          updated_at?: string | null
        }
        Update: {
          city?: string | null
          created_at?: string | null
          employer_id?: number | null
          post_code?: string | null
          project_type?: string | null
          site_id?: number
          site_name?: string | null
          state?: string | null
          street_address?: string | null
          supervisor_name?: string | null
          supervisor_telephone?: string | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "workplace_locations_employer_id_fkey"
            columns: ["employer_id"]
            isOneToOne: false
            referencedRelation: "employers"
            referencedColumns: ["employer_id"]
          },
        ]
      }
// ...
      workers: {
        Row: {
          basis_of_employment: string | null
          country_of_birth: string | null
          created_at: string | null
          date_of_birth: string | null
          email: string | null
          employer_id: number | null
          employment_arrangement: string | null
          employment_type: string | null
          family_name: string | null
          gender: string | null
          given_name: string | null
          interpreter_required: boolean | null
          mobile_number: string | null
          occupation: string | null
          phone_number: string | null
          post_code: string | null
          preferred_language: string | null
          residential_address: string | null
          state: string | null
          suburb: string | null
          updated_at: string | null
          worker_id: number
        }
        Insert: {
          basis_of_employment?: string | null
          country_of_birth?: string | null
          created_at?: string | null
          date_of_birth?: string | null
          email?: string | null
          employer_id?: number | null
          employment_arrangement?: string | null
          employment_type?: string | null
          family_name?: string | null
          gender?: string | null
          given_name?: string | null
          interpreter_required?: boolean | null
          mobile_number?: string | null
          occupation?: string | null
          phone_number?: string | null
          post_code?: string | null
          preferred_language?: string | null
          residential_address?: string | null
          state?: string | null
          suburb?: string | null
          updated_at?: string | null
          worker_id?: number
        }
        Update: {
          basis_of_employment?: string | null
          country_of_birth?: string | null
          created_at?: string | null
          date_of_birth?: string | null
          email?: string | null
          employer_id?: number | null
          employment_arrangement?: string | null
          employment_type?: string | null
          family_name?: string | null
          gender?: string | null
          given_name?: string | null
          interpreter_required?: boolean | null
          mobile_number?: string | null
          occupation?: string | null
          phone_number?: string | null
          post_code?: string | null
          preferred_language?: string | null
          residential_address?: string | null
          state?: string | null
          suburb?: string | null
          updated_at?: string | null
          worker_id?: number
        }
        Relationships: [
          {
            foreignKeyName: "workers_employer_id_fkey"
            columns: ["employer_id"]
            isOneToOne: false
            referencedRelation: "employers"
            referencedColumns: ["employer_id"]
          },
        ]
      }
    }
// ...
  }
}